requirements.txt
requests
beautifulsoup4
lxml
rapidfuzz
pillow
fastapi
uvicorn[standard]
python-multipart
boto3
